+++ packages/orq-rc/src/funcs/deploymentsStream.ts
@@ -107,7 +107,7 @@ async function $do(
   const context = {
     baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "DeploymentStream",
-    oAuth2Scopes: [],
+    oAuth2Scopes: null,
 
     resolvedSecurity: requestSecurity,