+++ packages/orq-rc/src/models/operations/getallprompts.ts
@@ -422,6 +422,9 @@ export const GetAllPromptsUseCases = {
   Sql: "SQL",
   Summarization: "Summarization",
   Tagging: "Tagging",
+  TranslationDocument: "Translation (document)",
+  TranslationSentences: "Translation (sentences)",
+  SentimentAnalysis: "Sentiment analysis",
 } as const;
 export type GetAllPromptsUseCases = ClosedEnum<typeof GetAllPromptsUseCases>;