+++ packages/orq-rc/src/models/operations/getpromptversion.ts
@@ -412,6 +412,9 @@ export const GetPromptVersionUseCases = {
   Sql: "SQL",
   Summarization: "Summarization",
   Tagging: "Tagging",
+  TranslationDocument: "Translation (document)",
+  TranslationSentences: "Translation (sentences)",
+  SentimentAnalysis: "Sentiment analysis",
 } as const;
 export type GetPromptVersionUseCases = ClosedEnum<
   typeof GetPromptVersionUseCases