+++ packages/orq-rc/src/sdk/sdk.ts
@@ -8,6 +8,7 @@ import { Datasets } from "./datasets.js";
 import { Deployments } from "./deployments.js";
 import { Feedback } from "./feedback.js";
 import { Files } from "./files.js";
+import { Knowledge } from "./knowledge.js";
 import { Prompts } from "./prompts.js";
 import { Remoteconfigs } from "./remoteconfigs.js";
 
@@ -45,5 +46,10 @@ export class Orq extends ClientSDK {
   private _datasets?: Datasets;
   get datasets(): Datasets {
     return (this._datasets ??= new Datasets(this._options));
+  }
+
+  private _knowledge?: Knowledge;
+  get knowledge(): Knowledge {
+    return (this._knowledge ??= new Knowledge(this._options));
   }
 }