+++ packages/orq-rc/RELEASES.md
@@ -1089,3 +1089,13 @@ Based on:
 - [typescript v3.3.0-rc.11] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.11 - packages/orq-rc
+
+## 2025-03-27 14:20:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.525.0 (2.562.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.0-rc.12] packages/orq-rc
+### Releases
+- [NPM v3.3.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.12 - packages/orq-rc